chatservices/requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic
langchain-fireworks
langchain-core
langchain-cli
chromadb
chromadb-client
docx2txt
pypdf
streamlit
tiktoken
unstructured
python-pptx
pdfminer
pdf2image
tesseract
unstructured_inference
pdfminer.six
pinecone-client
python-dotenv
sentence_transformers
pdfplumber
beautifulsoup4
PyMuPDF
reportlab